backend/requirements.txt
Flask
flask-cors
python-pptx
pandas
